+++ linbox/linbox/vector/reverse.h
@@ -3,20 +3,20 @@
  *
  * ------------------------------------
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library LinBox.
- * 
+ *
  * LinBox is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -80,8 +80,8 @@ namespace LinBox
 
 		// Element access
 
-		inline reference       operator[] (size_type n)       { return (begin ())[n]; }
-		inline const_reference operator[] (size_type n) const { return (begin ())[n]; }
+		inline reference       operator[] (size_type n)       { return ((begin ())[(ptrdiff_t)n]); }
+		inline const_reference operator[] (size_type n) const { return ((begin ())[(ptrdiff_t)n]); }
 
 		// the method "at" does appear to be implemented
 		// in the gnu implementation of the STL
